requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pypdf
python-docx
openai
chromadb
fastembed

# для scripts/fetch_handbook_links.py (парсинг ссылок из handbook)
requests
beautifulsoup4

